+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.43
 
 * bug fixed for `SequentialIssuer`
+* added `TerminableThread.safe_wait_condition`